Creating GenreralEntiy and generalizing the factory
GeneralEntity is the super class for all the beans .All the entities will have  3 vectors    that will store the names of the fields , the labels in front of them and their types.

In the class of the bean , the constructor will create the entity by initializing all the elements of the vectors.

The factorry is updated .
The method that creates the views has an argument, if the  argumnet is the one that we need   then the correct  class is made.

The labels are created with a  for and added to the container.
If the current field is a text field ,it is created and and added to the continer.